+++ esmerald/conf/global_settings.py
@@ -1397,6 +1397,8 @@ class EsmeraldAPISettings(BaseSettings):
             def is_type(self, value: Any) -> bool:
                 return has(value)
 
+            is_type_structure = is_type
+
             def serialize(self, obj: Any) -> Any:
                 return asdict(obj)