+++ esmerald/responses/encoders.py
@@ -1,7 +1,10 @@
-from typing import Any
+from functools import partial
+from inspect import isclass
+from typing import Any, cast
 
 import orjson
 
+from esmerald.encoders import ENCODER_TYPES_CTX, json_encode
 from esmerald.responses.json import BaseJSONResponse
 
 try:
@@ -17,15 +20,48 @@ class ORJSONResponse(BaseJSONResponse):
     In the same way the JSONResponse is used, so is the `ORJSONResponse`.
     """
 
+    @staticmethod
+    def transform(value: Any, encoders: Any = None) -> dict[str, Any]:
+        """
+        The transformation of the data being returned.
+
+        Supports all the default encoders from Lilya and custom from Esmerald.
+        """
+        return cast(
+            dict[str, Any],
+            json_encode(
+                value,
+                json_encode_fn=orjson.dumps,
+                post_transform_fn=orjson.loads,
+                with_encoders=encoders,
+            ),
+        )
+
     def make_response(self, content: Any) -> bytes:
-        return orjson.dumps(
-            content,
-            default=self.transform,
-            option=orjson.OPT_SERIALIZE_NUMPY | orjson.OPT_OMIT_MICROSECONDS,
+        encoders = (
+            (
+                (
+                    *(encoder() if isclass(encoder) else encoder for encoder in self.encoders),
+                    *ENCODER_TYPES_CTX.get(),
+                )
+            )
+            if self.encoders
+            else None
+        )
+        return cast(
+            bytes,
+            json_encode(
+                content,
+                json_encode_fn=partial(
+                    orjson.dumps, option=orjson.OPT_SERIALIZE_NUMPY | orjson.OPT_OMIT_MICROSECONDS
+                ),
+                post_transform_fn=None,
+                with_encoders=encoders,
+            ),
         )
 
 
-class UJSONResponse(BaseJSONResponse):
+class UJSONResponse(ORJSONResponse):
     """
     An alternative to `JSONResponse` and performance wise, faster.
 
@@ -34,7 +70,11 @@ class UJSONResponse(BaseJSONResponse):
 
     def make_response(self, content: Any) -> bytes:
         assert ujson is not None, "You must install the encoders or ujson to use UJSONResponse"
-        return ujson.dumps(content, ensure_ascii=False).encode("utf-8")
+        # UJSON is actually in maintainance mode only, so use mostly orjson
+        # https://github.com/ultrajson/ultrajson
+        return ujson.dumps(self.transform(content, self.encoders), ensure_ascii=False).encode(
+            "utf-8"
+        )
 
 
 __all__ = ["ORJSONResponse", "UJSONResponse"]