+++ tests/dependencies/test_simple_case_injected_annotation.py
@@ -1,14 +1,11 @@
-from typing import TYPE_CHECKING, List
+from typing import List
 
 from esmerald import post
 from esmerald.injector import Inject
 from esmerald.routing.apis.views import APIView
 from esmerald.routing.gateways import Gateway
 from esmerald.testclient import create_client
-from tests.dependencies.samples import DocumentService
-
-if TYPE_CHECKING:
-    from tests.dependencies.samples import DocumentCreateDTO
+from tests.dependencies.samples import DocumentCreateDTO, DocumentService
 
 
 class DocumentAPIView(APIView):
@@ -25,7 +22,6 @@ class DocumentAPIView(APIView):
 
 
 def test_injection():
-
     with create_client(routes=[Gateway(handler=DocumentAPIView)]) as client:
         response = client.post("/", json={"name": "test", "content": "test"})
         assert response.status_code == 201